+++ src/dom_entity_names.rs
@@ -1,3 +1,4 @@
 pub mod attributes;
+pub mod events;
 pub mod styles;
 pub mod tags;
