+++ src/dom_entity_names/events.rs
@@ -1,0 +1,67 @@
+/// Similar to tag population.
+macro_rules! make_events {
+    // Create shortcut macros for any element; populate these functions in this module.
+    { $($event_camel:ident => $event:expr),+ } => {
+
+        /// The Ev enum restricts element-creation to only valid event names, as defined here:
+        /// [https://developer.mozilla.org/en-US/docs/Web/Evs](https://developer.mozilla.org/en-US/docs/Web/Evs)
+        #[derive(Clone, Copy, Debug, PartialEq)]
+        pub enum Ev {
+            $(
+                $event_camel,
+            )+
+        }
+
+        impl Ev {
+            pub fn as_str(&self) -> &str {
+                match self {
+                    $ (
+                        Ev::$event_camel => $event,
+                    ) +
+                }
+            }
+        }
+
+        impl From<&str> for Ev {
+            fn from(event: &str) -> Self {
+                match event {
+                    $ (
+                          $event => Ev::$event_camel,
+                    ) +
+                    _ => {
+                        crate::error(&format!("Can't find this event: {}", event));
+                        Ev::Click
+                    }
+                }
+            }
+        }
+
+        impl From<String> for Ev {
+            fn from(event: String) -> Self {
+                match event.as_ref(){
+                    $ (
+                          $event => Ev::$event_camel,
+                    ) +
+                    _ => {
+                        crate::error(&format!("Can't find this event: {}", event));
+                        Ev::Click
+                    }
+                }
+            }
+        }
+
+        impl ToString for Ev {
+            fn to_string( & self ) -> String {
+                match self {
+                    $ (
+                        Ev::$ event_camel => $ event.into(),
+                    ) +
+
+                }
+            }
+        }
+    }
+}
+
+mod event_names;
+pub use event_names::Ev;
