+++ src/dom_entity_names/events/event_names.rs
@@ -1,0 +1,45 @@
+// Comprehensive list: https://developer.mozilla.org/en-US/docs/Web/Events
+make_events! {
+    Cached => "cached", Error => "error", Abort => "abort", Load => "load", BeforeUnload => "beforeunload",
+    Unload => "unload", Online => "online", Offline => "offline", Focus => "focus", Blur => "blur",
+    Open => "open", Message => "message", Close => "close", PageHide => "pagehide",
+    PageShow => "pageshow", PopState => "popstate", AnimationStart => "animationstart", AnimationEnd => "animationend",
+    AnimationIteration => "animationiteration", TransitionStart => "transtionstart", TransitionEnd => "transitionend",
+    TranstionRun => "transitionrun",
+
+    Rest => "rest", Submit => "submit", BeforePrint => "beforeprint", AfterPrint => "afterprint",
+    CompositionStart => "compositionstart", CompositionUpdate => "compositionupdate", CompositionEnd => "compositionend",
+
+    FullScreenChange => "fullscreenchange", FullScreenError => "fullscreenerror", Resize => "resize",
+    Scroll => "scroll", Cut => "cut", Copy => "copy", Paste => "paste",
+
+    KeyDown => "keydown", KeyUp => "keyup",
+    KeyPress => "keypress", AuxClick => "auxclick", Click => "click", ContextMenu => "contextmenu", DblClick => "dblclick",
+    MouseDown => "mousedown", MouseEnter => "mouseenter", MouseLeave => "mouseleave",
+    MouseMove => "mousemove", MouseOver => "mouseover", MouseOut => "mouseout", MouseUp => "mouseup",
+    PointerLockChange => "pointerlockchange", PointerLockError => "pointerlockerror", Select => "select",
+    Wheel => "wheel",
+
+    PointerOver => "pointerover", PointerEnter => "pointerenter",
+    PointerDown => "pointerdown", PointerMove => "pointermove", PointerUp => "pointerup",
+    PointerCancel => "pointercancel", PointerOut => "pointerout", PointerLeave => "pointerleave",
+    GotPointerCapture => "gotpointercapture", LostPointerCapture => "lostpointercapture",
+
+    Drag => "drag", DragEnd => "dragend", DragEnter => "dragenter", DragStart => "dragstart", DragLeave => "dragleave",
+    DragOver => "dragover", Drop => "drop",
+
+    AudioProcess => "audioprocess", CanPlay => "canplay", CanPlayThrough => "canplaythrough", Complete => "complete",
+    DurationChange => "durationchange", Emptied => "emptied", Ended => "ended", LoadedData => "loadeddata",
+    LoadedMetaData => "loadedmetadata", Pause => "pause", Play => "play", Playing => "playing", RateChange => "ratechange",
+    Seeked => "seeked", Seeking => "seeking", Stalled => "stalled", Suspend => "suspend", TimeUpdate => "timeupdate",
+    VolumeChange => "volumechange",
+
+    // todo finish this
+
+    Change => "change",
+
+    Input => "input",
+
+    // deprecated, @TODO: remove
+    TriggerUpdate => "triggerupdate"
+}
